Change the modify directive to use element form. (#23)
This allows use of CDATA sections to make entering regexes a bit easier.   Also added expand attribute as a way to tell Plaster to expand the string or not.  This would be like selecting between single quotes (not expanded) which is the default or double quotes (expanded).

There is an issue with ExpandString that needs to be resolved.  Right now if you expand the substitute string and want to use regex $1 you have to specify it as ``$1.  It seems to me it should be only `$1.