Fixes #55 - misc issues.
This removes the UnrecognizedAttribute_F2 msg, replaces a literal 'Create' string with the msg table entry for Create, fixes Test-PlasterManifest to handle a missing target namespace and emits the specific reason why a not well-formed XML file is not well-formed.  Also, added a -NoLogo parameter to Invoke-Plaster to cut down on the noise when running the Pester tests.  And added a Pester test for the correct doc element name but missing a target namespace.