Fix #38 - localization now finds fr-FR resources (#64)
The main trick, according to the docs, is to have Import-LocalizedData ignore the error when the local culture can't be found in the module resources.  The the data section (with the fallback strings) is used.  One point, I don't like this duplication between en-US and the data section in Plaster.psm1.  This is not DRY.  What I want is a way to declare the modules default resources as en-US.  Is there no way to do this?