Modify now uses conflict res (#75)
This was a bit of a significant change and caused me to add several features.  When a file is modified that the template added previously, the operation is now called "Update" and  will not result in a conflict.  If the file pre-existed the application of the template then normal file conflict detect/res comes into play.  Also changed color of "intermediate" ops like Expanding and Modifying to display in Magenta to distinguish them from the Green of create.  Expand and modify now do their work in temp files. Then temp files are then passed to the file conflict detect/res script.

Without these changes the combo on <file source="tasks.json"..> and <modify path="tasks.json"..> would result in a conflict that need to be answered during a template invoke on an empy dest directory. Clearly this was not desirable.