Fix InstallPath issue in build.psake.ps1
This change fixes an issue which causes the build.psake.ps1 script to fail
when run in a host that does not set the $profile variable.  Since the
InstallPath property had a default value which uses $profile directly,
this would cause an error to be thrown when a null $profile is
encountered, preventing the build script from continuing.

The fix is to use $null for default and then construct the default value
and resolve the default path inside of the Install task.  This fix has
also been applied to the example NewModuleTemplate which follows a similar
structure.