Add TestOutputFile and TestOutputType properties to psake script
This change adds two new configurable properties to the build.psake.ps1
script so that a test output file can be created when Pester tests are
run.  This is useful in contexts like AppVeyor where this output file can
be used to show test results in the CI interface.

These changes have also been applied to the build.psake.ps1 file in the
example NewModuleTemplate.