Add appveyor.yml to enable CI builds (#125)
* Fix InstallPath issue in build.psake.ps1

This change fixes an issue which causes the build.psake.ps1 script to fail
when run in a host that does not set the $profile variable.  Since the
InstallPath property had a default value which uses $profile directly,
this would cause an error to be thrown when a null $profile is
encountered, preventing the build script from continuing.

The fix is to use $null for default and then construct the default value
and resolve the default path inside of the Install task.  This fix has
also been applied to the example NewModuleTemplate which follows a similar
structure.

* Add TestOutputFile and TestOutputType properties to psake script

This change adds two new configurable properties to the build.psake.ps1
script so that a test output file can be created when Pester tests are
run.  This is useful in contexts like AppVeyor where this output file can
be used to show test results in the CI interface.

These changes have also been applied to the build.psake.ps1 file in the
example NewModuleTemplate.

* Add appveyor.yml to enable CI builds

This change adds an appveyor.yml configuration file to enable CI builds
for the Plaster project.

Resolves #22.

* Fix Pester test issues causing failures on AppVeyor

This change fixes a few issues in Plaster's Pester tests which cause
failures on AppVeyor.  Specifically, there are issues with module loading
and manifest file comparisons due to AppVeyor's environment.