Don't attempt XML schema validation on .NET Core.
Fix #107.  This approach checks to see if the type System.Xml.Schema.XmlSchemaSet is available.  Right *now* on .NET Core it is not.  In this case, we don't execute the script that configures the reader to fire validation events.  In the future, when .NET Core supports netstandard 2.0, the validation code should just start working without any changes required.