Add plasterVersion to manifest schema.
The plasterVersion is usefully when a newer version of Plaster supports a new runtime behavior (typically new commands and/or variables in the constrained runspace) and a template needs that behavior.  The presence of this attribute will cause Test-PlasterManifest (and Invoke-PlasterTemplate) to validate that the Plaster version is high enough.  If not, an error message is emitted telling the user to update the Plaster module.

Also added a new constrained runspace variable PLASTER_Version for runtime checks.  This can be used when the template can fallback gracefully depending on the version of Plaster.