Remove condition attribute on choice.
Not ready to add this yet.  More works needs to be done to figure out how to make the <parameter default attribute work ocrrectly in face of dynamic list of choices.

Add Pester test for parameter condition attribute.
Fix Test task in tasks.json by temporarily replacing ${env.windir} with C:\\Windows. This seems to be a regression in VSCode.